New proto support and folder organization
